[userspace] [coherent dedispersion] correct the coefficient in formula
The unit of frequency in MHz, not Hz, so additional 1*10^6 coefficient
is required.

Moreover, precision of float is not enough here because `phase` can up
to 10^9 while difference between two frequency channels is ~1-10 .
Forcing `srtb::real` to `double` should fix this, but due to limited
VRAM, will try more solutions on this.

This has been bothering me for over a month ...