[userspace] baseband input size unit: bytes -> count
Almost all other part of this code uses data point count as length of
a group of data, instead the actual memory size (in bytes) .
So for code consistency this length is re-defined.

Also changed the behaviour of srtb::unpack to align with this change.

File reader is also changed.