[userspace] pipe: split ifft and refft pipe
Current way to reserve samples (because of coherent dedispersion) and
keep data continuity is to change refft's batch_size.
Prior changes to nsamps_reserved() relate to this.

Ring buffer method is commented out and will be cleaned next clean up.
It' s kept here in case needed in the future.

Also updated doc.