[userspace] gui: add spectrum thumbnail generation
This method just generates thumbnail from the spectrum,
size of the thumbnail should be the same as pixmap on GUI.
Also dramatically simplifies Spectrum Image Provider.
Notice, in this method waterfall will no longer available;
but as `watfft_1d_c2c` is going to be used,
direction of the generated spectrum won't allow next pipes to
implement waterfall effect easily, so `drop()`.

Signed-off-by: fxzjshm <fxzjshm@163.com>