io: remove stub of rdma
No FPGA currently supports this as of my knowledge. It's hard.

Also removing the long-existed "[userspace]" tag in commit message,
since with the experience of UCX, no kernel module need to be written;
and if any input requires kernel module/driver (e.g. PCIe sample cards),
it is unlikely that those code will be in this repo.

Signed-off-by: fxzjshm <fxzjshm@163.com>